Fixed block type dropdown API and small refactor
