+++ packages/react/src/components/FormattingToolbar/mantine/DefaultButtons/NestBlockButtons.tsx
@@ -9,7 +9,7 @@ import { RiIndentDecrease, RiIndentIncrease } from "react-icons/ri";
 
 import { useBlockNoteEditor } from "../../../../editor/BlockNoteContext";
 import { useEditorContentOrSelectionChange } from "../../../../hooks/useEditorContentOrSelectionChange";
-import { ToolbarButton } from "../../../../components-shared/Toolbar/ToolbarButton";
+import { ToolbarButton } from "../../../mantine-shared/Toolbar/ToolbarButton";
 
 export const NestBlockButton = () => {
   const editor = useBlockNoteEditor<
