+++ packages/react/src/components/FormattingToolbar/mantine/DefaultButtons/ReplaceImageButton.tsx
@@ -10,7 +10,7 @@ import { RiImageEditFill } from "react-icons/ri";
 
 import { useBlockNoteEditor } from "../../../../editor/BlockNoteContext";
 import { useSelectedBlocks } from "../../../../hooks/useSelectedBlocks";
-import { ToolbarButton } from "../../../../components-shared/Toolbar/ToolbarButton";
+import { ToolbarButton } from "../../../mantine-shared/Toolbar/ToolbarButton";
 import { ImageToolbar } from "../../../ImageToolbar/mantine/ImageToolbar";
 
 export const ReplaceImageButton = () => {
