+++ packages/react/src/components/mantine-shared/ColorPicker/ColorIcon.tsx
