+++ packages/react/src/components/mantine-shared/Toolbar/Toolbar.tsx
@@ -2,8 +2,7 @@ import { mergeCSSClasses } from "@blocknote/core";
 import { Group } from "@mantine/core";
 import { forwardRef, HTMLAttributes } from "react";
 
-// TODO: move to mantine dir?
-export const ToolbarWrapper = forwardRef<
+export const Toolbar = forwardRef<
   HTMLDivElement,
   HTMLAttributes<HTMLDivElement>
 >((props, ref) => {
